pyproject.toml
[project]
authors = [
  {name = "K8s Admin Team"},
]
description = "AI-powered Kubernetes cluster administrator using MCP Server"
license = {text = "Apache-2.0"}
name = "k8s-admin-agent"
readme = "README.md"
requires-python = ">=3.13,<3.15"
version = "1.0.0"

dependencies = [
  "beeai-framework[a2a]>=0.1.78",
  "tiktoken>=0.12.0",
  "fastuuid>=0.14.0",
  "agentstack-sdk",
  "openinference-instrumentation-beeai>=0.1.14",
  "pydantic-settings>=2.9.0",
  "uvicorn>=0.37.0,<0.38.0",
  "httpx>=0.27.0",
  "loguru>=0.7.0",
  "pyyaml>=6.0",
]

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120

[project.scripts]
server = "k8s_admin_agent.agent:serve"

[tool.pyrefly]
project-includes = [
  "**/*.py*",
  "**/*.ipynb",
]

[build-system]
build-backend = "uv_build"
requires = ["uv_build>=0.10.0,<0.11.0"]

[dependency-groups]
dev = [
  "pyrefly>=0.52.0",
  "watchfiles>=1.1.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]
